Enabled HALF for fill() and zero() methods. Moved them into THTensorFill
