+++ aten/src/TH/THTensor.h
@@ -25,6 +25,13 @@
 #include <TH/generic/THTensorMath.h>
 #include <TH/THGenerateAllTypes.h>
 
+/* fill and zero*/
+#include <TH/generic/THTensorFill.h>
+#include <TH/THGenerateAllTypes.h>
+
+#include <TH/generic/THTensorFill.h>
+#include <TH/THGenerateHalfType.h>
+
 /* convolutions */
 #include <TH/generic/THTensorConv.h>
 #include <TH/THGenerateAllTypes.h>
