+++ aten/src/TH/THTensorFill.cpp
@@ -1,0 +1,8 @@
+#include <TH/THTensor.hpp>
+#include <TH/THVector.h>
+
+#include <TH/generic/THTensorFill.cpp>
+#include <TH/THGenerateAllTypes.h>
+
+#include <TH/generic/THTensorFill.cpp>
+#include <TH/THGenerateHalfType.h>
