+++ aten/src/TH/THVector.cpp
@@ -21,5 +21,11 @@
 #include <TH/generic/THVectorDefault.cpp>
 #include <TH/THGenerateAllTypes.h>
 
+#include <TH/generic/THVectorDefault.cpp>
+#include <TH/THGenerateHalfType.h>
+
 #include <TH/generic/THVectorDispatch.cpp>
 #include <TH/THGenerateAllTypes.h>
+
+#include <TH/generic/THVectorDispatch.cpp>
+#include <TH/THGenerateHalfType.h>
